update securiter for security work in progress

+++ src/main/java/Mboussaid/laFactureFacile/Repository/JwtRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import Mboussaid.laFactureFacile.Models.Jwt;
 
-public interface JwtRepository extends JpaRepository<Jwt, Long> {
+public interface JwtRepository extends JpaRepository<Jwt, Integer> {
     Optional<Jwt> findByValueAndIsExpiredAndIsBlackListed(String value, boolean isExpired, boolean isBlackListed);
 
     @Query("FROM Jwt j WHERE j.user.email = :email AND j.isExpired = :isExpired AND j.isBlackListed = :isBlackListed")
